actually validate authTag
The authTag is used to make sure that the data hasn't been tampered with
/ corrupted: https://crypto.stackexchange.com/a/101716

Before this commit, data could be corrupted and decryption would
"succeed", but produce bad data. I think it's probably a good call to
explicitly fail in this case, but I could see someone wanting a way to
disable this to do at least a partial recovery of the deadbolted files.

Maybe I will have to add a config option for this.